some changes for websocket config -> not yet fixed
