+++ samples/core/src/main/resources/application.properties
