+++ samples/core/src/main/resources/config/app/gui/dispatcher-servlet.xml
@@ -6,7 +6,6 @@
     http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
 
   <context:component-scan base-package="io.oasp.gastronomy.restaurant.general.gui.api"/>
-  <context:component-scan base-package="io.oasp.gastronomy.restaurant.*.websocket"/>
 
   <beans:bean class="org.springframework.web.servlet.view.UrlBasedViewResolver">
     <beans:property name="viewClass" value="org.springframework.web.servlet.view.JstlView"/>
