+++ samples/core/src/main/resources/config/application.properties
@@ -12,3 +12,7 @@ flyway.clean-on-validation-error=true
 logging.level.org.springframework.web=DEBUG
 logging.level.org.hibernate=ERROR
 
+# Flyway for setting up testdata in database
+#database.migration.auto=true
+#database.migration.testdata=false
+#database.migration.clean=false
