+++ samples/core/src/main/resources/db/migration/h2/V0001__R001_Create_schema.sql
@@ -129,3 +129,4 @@ CREATE TABLE PUBLIC.REVINFO(
     timestamp BIGINT NOT NULL,
     user VARCHAR(255)
 );
+
