+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/salesmanagement/batch/impl/billexport/BillExportJobTest.java
@@ -1,5 +1,8 @@
 package io.oasp.gastronomy.restaurant.salesmanagement.batch.impl.billexport;
 
+import io.oasp.gastronomy.restaurant.SpringBootApp;
+import io.oasp.gastronomy.restaurant.general.common.AbstractSpringBatchIntegrationTest;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -7,7 +10,6 @@ import java.io.IOException;
 
 import javax.inject.Inject;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,19 +19,17 @@ import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.test.AssertFile;
+import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.core.io.ClassPathResource;
-import org.springframework.test.context.ContextConfiguration;
-
-import io.oasp.gastronomy.restaurant.general.common.AbstractSpringBatchIntegrationTest;
-import io.oasp.module.configuration.common.api.ApplicationConfigurationConstants;
+import org.springframework.test.context.web.WebAppConfiguration;
 
 /**
  * End-To-End test job "import offer management from csv"
  *
  * @author jczas
  */
-@Ignore
-@ContextConfiguration(locations = { ApplicationConfigurationConstants.BEANS_BATCH })
+@SpringApplicationConfiguration(classes = { SpringBootApp.class }, locations = { "classpath:/config/app/batch/beans-batch.xml" })
+@WebAppConfiguration
 public class BillExportJobTest extends AbstractSpringBatchIntegrationTest {
   private static final Logger LOG = LoggerFactory.getLogger(AbstractSpringBatchIntegrationTest.class);
