+++ samples/core/src/test/resources/config/application.properties
@@ -1,3 +1,8 @@
 # Flyway for Database Setup and Migrations
 flyway.enabled=true
 flyway.clean-on-validation-error=true
+
+# Flyway for setting up testdata in database
+database.migration.auto=true
+database.migration.testdata=false
+database.migration.clean=false
